Move all exceptions into 1 package
