+++ src/main/java/com/refactoringflow/refactoringflowbackend/controller/AssignmentController.java
@@ -1,6 +1,6 @@
 package com.refactoringflow.refactoringflowbackend.controller;
 
-import com.refactoringflow.refactoringflowbackend.exception.ResourceNotFoundException;
+import com.refactoringflow.refactoringflowbackend.error.exceptions.ResourceNotFoundException;
 import com.refactoringflow.refactoringflowbackend.model.assignment.Assignment;
 import com.refactoringflow.refactoringflowbackend.model.user.Student;
 import com.refactoringflow.refactoringflowbackend.repository.AssignmentRepository;
