+++ src/main/java/com/refactoringflow/refactoringflowbackend/error/exceptions/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package com.refactoringflow.refactoringflowbackend.exception;
+package com.refactoringflow.refactoringflowbackend.error.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
